+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+- Zest comment including the extension version when recording.
+
 ## 0.0.12 - 2025-05-05
 
 ### Changed